requirements.txt
django
pillow
psycopg2-binary
djangorestframework
python-dotenv
django_filter